EX4/fullstack-app/web/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
redis

